# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Titre de votre API",
        "description": "Description de votre API",
        "contact": {
            "email": "support@votreapi.com"
        },
        "license": {
            "name": "Licence de votre API",
            "url": "URL de la licence"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Obtenir la liste des articles",
                "description": "Retourne la liste des articles avec des filtres optionnels.",
                "operationId": "2bb6d18420d6929d1d4175ace6521d21",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Inclure des relations",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "disponible",
                        "in": "query",
                        "description": "Filtrer les articles par disponibilité (oui/non)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des articles retournée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "ArticleResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1articles": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer un nouvel article",
                "description": "Ajoute un nouvel article à la base de données.",
                "operationId": "de780de04405bfc1ffb83cce99f65493",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "ArticleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Afficher un article par ID",
                "description": "Retourne un article spécifique par son ID.",
                "operationId": "bf9724a6309b55de3ed51502e1a15814",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retourné avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Mettre à jour un article",
                "description": "Met à jour un article existant.",
                "operationId": "3567f8e5d2eb2e799f139c75513d7283",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "UpdateArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "ArticleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "description": "Supprime un article en utilisant la suppression douce (soft delete).",
                "operationId": "51d3b08d51f0416b640aa7728b78bdb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article supprimé avec succès"
                    },
                    "404": {
                        "description": "Article non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Articles"
                ],
                "summary": "Mettre à jour le stock d'un article par ID",
                "description": "Met à jour la quantité de stock d'un article spécifique.",
                "operationId": "cf2eb9ab0d0680af30a7e92555ba9efa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "UpdateStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé"
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/articles/stock": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Mettre à jour les stocks des articles",
                "description": "Met à jour les quantités de stock pour plusieurs articles.",
                "operationId": "bebeaf27d5be355373f405fd37832575",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "articles": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "UpdateStockRequest"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stocks mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updatedArticles": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "quantite": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "notFoundArticles": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "quantite": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "En-tête d'autorisation JWT utilisant le schéma Bearer."
            }
        }
    },
    "security": [
        {
            "BearerToken": []
        }
    ]
}